code cleanup
git-svn-id: http://svn.miranda-ng.org/main/trunk@12120 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Tox/src/tox_profile.cpp
@@ -41,7 +41,7 @@ bool CToxProto::LoadToxProfile()
 	}
 
 	uint8_t *data = (uint8_t*)mir_calloc(size);
-	if (fread((char*)data, sizeof(char), size, profile) != size)
+	if (fread((char*)data, sizeof(char), size, profile) != (size_t)size)
 	{
 		fclose(profile);
 		debugLogA("CToxProto::LoadToxData: could not read tox profile");
